address comments and remove gemma specific code out of base file
